refactor: deduplicate audio_utils tests into dedicated test file [AI]
Addresses Gemini Code Assist review feedback on PR #171.

Changes:
- Created tests/text_to_audio/test_audio_utils.py with all tests for
  is_valid_wav() and wrap_pcm_in_wav() functions
- Removed duplicate WrapPcmInWavTest, IsValidWavTest, and
  WrapPcmEdgeCasesTest classes from test_google_tts_provider.py
- Removed duplicate GeminiWrapPcmInWavTest class from
  test_gemini_tts_provider.py
- All test methods now follow proper snake_case convention

Benefits:
- Single source of truth for audio_utils tests
- Follows DRY principle - no more duplicate tests
- Improves test maintainability and organization
- Tests are now organized by the module they test

Stats:
- Removed 198 lines of duplicate test code
- Created 104-line dedicated test file for audio_utils
- Net reduction: 94 lines
